requirements.txt
rich
pytest
requests
groq
langchain
sentence-transformers
faiss-cpu
python-dotenv
pydantic
typer
fastapi
uvicorn
loguru
gitpython
unidiff